Make static mut work by moving .got section
It seems to depend on the location of the .got section, even though
everything is supposed to be statically linked. I was mistaken that it
was going to be written to at some point so it had to be in RAM, but
that seems to be in the case of dynamic linking.